Time: 3 ms (98.36%), Space: 10.9 MB (72.38%) - LeetHub

+++ 1656-count-good-triplets/1656-count-good-triplets.cpp
@@ -3,16 +3,16 @@ public:
     int countGoodTriplets(vector<int>& arr, int a, int b, int c) 
     {
         int answer = 0;
-        int arrSize = arr.size();
+        unsigned short arrSize = arr.size();
 
-        for (int i = 0; i < arrSize - 2; ++i)
+        for (unsigned short i = 0; i < arrSize - 2; ++i)
         {
-            for (int j = i + 1; j < arrSize - 1; ++j)
+            for (unsigned short j = i + 1; j < arrSize - 1; ++j)
             {
                 if (abs(arr[i] - arr[j]) > a)
                     continue;
 
-                for (int k = j + 1; k < arrSize; ++k)
+                for (unsigned short k = j + 1; k < arrSize; ++k)
                     if (abs(arr[j] - arr[k]) <= b && abs(arr[i] - arr[k]) <= c)
                         ++answer;
             }
